requirements.txt
flask
langchain-community
openai
beautifulsoup4
requests
numpy
faiss-cpu
python-dotenv
tiktoken
